+++ src/main/java/com/woowacourse/dsgram/web/controller/UserApiController.java
@@ -6,6 +6,7 @@ import com.woowacourse.dsgram.service.dto.user.LoginUserDto;
 import com.woowacourse.dsgram.service.dto.user.SignUpUserDto;
 import com.woowacourse.dsgram.service.dto.user.UserDto;
 import com.woowacourse.dsgram.web.argumentresolver.UserSession;
+import com.woowacourse.dsgram.web.controller.exception.InvalidPatternException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
