+++ src/main/java/com/woowacourse/dsgram/web/controller/exception/InvalidPatternException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.dsgram.web.controller;
+package com.woowacourse.dsgram.web.controller.exception;
 
 public class InvalidPatternException extends RuntimeException {
     public InvalidPatternException(String message) {
